requirements.txt
accelerate==0.24.1
fastapi==0.104.1
numpy==1.26.2
pandas==2.1.3
protobuf==4.25.1
pydantic==2.5.2
python-dotenv==1.0.0
python-multipart==0.0.6
rainbow-bridge-logger==0.7.0
scikit-learn==1.3.2
scipy==1.11.4
torch==2.1.1
transformers==4.35.2
typing_extensions==4.8.0
uvicorn==0.24.0
requests==2.29.0